REG-15759 unify and update standlone scripts, add README
